// src/generated/barcoding-v3.ts
export interface components {
    schemas: {
        /** @description Barcode Response model containing a basic product information and the quantity. */
        ProductBarcodeResponse: {
            /**
             * @description The Digi-Key part number.
             * @example P189-ND
             */
            DigiKeyPartNumber?: string;
            /**
             * @description The manufacturer part number.
             * @example CR2032
             */
            ManufacturerPartNumber?: string;
            /**
             * @description Manufacturer of the product.
             * @example Panasonic - BSG
             */
            ManufacturerName?: string;
            /**
             * @description Catalog description of the product.
             * @example BATTERY LITHIUM 3V COIN 20MM
             */
            ProductDescription?: string;
            /**
             * Format: int32
             * @description Quantity of the product.
             * @example 5
             */
            Quantity?: number;
        };
        ApiValidationError: {
            Field?: string;
            Message?: string;
        };
        ApiErrorResponse: {
            readonly ErrorResponseVersion?: string;
            /** Format: int32 */
            StatusCode?: number;
            ErrorMessage?: string;
            ErrorDetails?: string;
            RequestId?: string;
            ValidationErrors?: components["schemas"]["ApiValidationError"][];
        };
        /** @description Data retrieved from a 2D barcode on a product. */
        Product2DBarcodeResponse: {
            /**
             * @description The Digi-Key part number.
             * @example P189-ND
             */
            DigiKeyPartNumber?: string;
            /**
             * @description The manufacturer part number.
             * @example CR2032
             */
            ManufacturerPartNumber?: string;
            /**
             * @description Manufacturer of the product.
             * @example Panasonic - BSG
             */
            ManufacturerName?: string;
            /**
             * @description Catalog description of the product.
             * @example BATTERY LITHIUM 3V COIN 20MM
             */
            ProductDescription?: string;
            /**
             * Format: int32
             * @description Quantity of the product.
             * @example 5
             */
            Quantity?: number;
            /**
             * Format: int32
             * @description The Id for the Salesorder. Note that if the barcode was from a "Prepack" label, this will not be populated.
             * @example 12345
             */
            SalesorderId?: number;
            /**
             * Format: int32
             * @description The Id for the invoice. Note that if the barcode was from a "Prepack" label, this will not be populated.
             * @example 12345
             */
            InvoiceId?: number;
            /**
             * @description Customer PurchaseOrder (custom notes or identifier) on the order. Note that if the barcode was from a "Prepack" label, this will not be populated.
             * @example April batch 1
             */
            PurchaseOrder?: string;
            /**
             * @description The country that the product orginates from.
             * @example China
             */
            CountryOfOrigin?: string;
            /** @description The Lot Code */
            LotCode?: string;
            /** @description The date code */
            DateCode?: string;
        };
        /** @description A detail from a Packlist containing the DigiKeyPartNumber and Quantity. */
        PackListDetail: {
            /**
             * @description The Digi-Key part number.
             * @example P189-ND
             */
            DigiKeyPartNumber?: string;
            /**
             * Format: int32
             * @description Quantity of the product.
             * @example 5
             */
            Quantity?: number;
        };
        /** @description Data retrieved from a Packlist. */
        PackListBarcodeResponse: {
            /**
             * Format: int32
             * @description The Id for the Salesorder.
             * @example 12345
             */
            SalesorderId?: number;
            /**
             * Format: int32
             * @description The Id for the invoice.
             * @example 12345
             */
            InvoiceId?: number;
            /**
             * Format: int32
             * @description PackList number.
             * @example 1
             */
            PackListNumber?: number;
            /** @description List of PackList Details. */
            PackListDetails?: components["schemas"]["PackListDetail"][];
        };
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}
export type $defs = Record<string, never>;
export interface operations {
    ProductBarcode: {
        parameters: {
            query?: {
                /**
                 * @description Comma separated list of fields to return. Used to customize response to reduce bandwidth by
                 *     selecting fields that you wish to receive. For example: "DigiKeyPartNumber,Quantity".
                 */
                includes?: string;
            };
            header: {
                /** @description OAuth Bearer Token. Please see<a href= "https://developer.digikey.com/documentation/oauth" target= "_blank" > OAuth 2.0 Documentation </a > page for more info. */
                Authorization: string;
                /** @description The Client Id for your App. */
                "X-DIGIKEY-Client-Id": string;
            };
            path: {
                /** @description Product barcode located on the product's anti-static bag. */
                barcode: string;
            };
            cookie?: never;
        };
        requestBody?: never;
        responses: {
            /** @description Success */
            200: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    "application/json": components["schemas"]["ProductBarcodeResponse"];
                };
            };
            /** @description Bad Request - The input model is invalid or malformed */
            400: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    "application/json": components["schemas"]["ApiErrorResponse"];
                };
            };
            /** @description Unauthorized - Token is expired or Client Id or subscription are invalid */
            401: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    "application/json": components["schemas"]["ApiErrorResponse"];
                };
            };
            /** @description Too Many Requests - Your rate limit has been exceeded */
            429: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    "application/json": components["schemas"]["ApiErrorResponse"];
                };
            };
            /** @description Service Unavailable - The service is temporarily unavailable */
            503: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    "application/json": components["schemas"]["ApiErrorResponse"];
                };
            };
        };
    };
    Product2DBarcode: {
        parameters: {
            query?: {
                /**
                 * @description Comma separated list of fields to return. Used to customize response to reduce bandwidth by
                 *     selecting fields that you wish to receive. For example: "DigiKeyPartNumber,Quantity"
                 */
                includes?: string;
            };
            header: {
                /** @description OAuth Bearer Token. Please see<a href= "https://developer.digikey.com/documentation/oauth" target= "_blank" > OAuth 2.0 Documentation </a > page for more info. */
                Authorization: string;
                /** @description The Client Id for your App. */
                "X-DIGIKEY-Client-Id": string;
            };
            path: {
                /**
                 * @description QR Code from a Digi-Key product label. The QR Code contains special ASCII symbols. These MUST be
                 *     encoded to be sent through the API. Please ensure the Record Separator character is encoded as \u241E and the Group
                 *     Separator character is encoded as \u241D
                 */
                barcode: string;
            };
            cookie?: never;
        };
        requestBody?: never;
        responses: {
            /** @description Success */
            200: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    "application/json": components["schemas"]["Product2DBarcodeResponse"];
                };
            };
            /** @description Bad Request - The input model is invalid or malformed */
            400: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    "application/json": components["schemas"]["ApiErrorResponse"];
                };
            };
            /** @description Unauthorized - Token is expired or Client Id or subscription are invalid */
            401: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    "application/json": components["schemas"]["ApiErrorResponse"];
                };
            };
            /** @description Too Many Requests - Your rate limit has been exceeded */
            429: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    "application/json": components["schemas"]["ApiErrorResponse"];
                };
            };
            /** @description Service Unavailable - The service is temporarily unavailable */
            503: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    "application/json": components["schemas"]["ApiErrorResponse"];
                };
            };
        };
    };
// ...
}