// src/generated/packing-list-v1.ts
export interface components {
    schemas: {
        DKProblemDetails: {
            type?: string;
            title?: string;
            /** Format: int32 */
            status?: number;
            detail?: string;
            instance?: string;
            correlationId?: string;
            readonly errors?: {
                [key: string]: string[];
            };
        };
        InvoicePackingList: {
            /** @description Item found from Invoice */
            PackingListItem?: components["schemas"]["PackingListItem"][];
            /**
             * Format: int32
             * @description The InvoiceID that you entered in the request
             */
            InvoiceIdEntered?: number;
            /** @description The value that you entered in the request. "True" is to return the encoded pdf string. "False" is to not. */
            IncludePdfEntered?: boolean;
            /** @description Pdf for this invoice has been found */
            PdfFound?: boolean;
            /** @description Enter this string in a string decoder to generate a pdf. */
            EncodedString?: string;
        };
        PackingListItem: {
            /**
             * Format: int32
             * @description Invoice ID of order. There is one per shipment.
             */
            InvoiceId?: number;
            /**
             * Format: int32
             * @description Sales order ID. There is one per order.
             */
            SalesOrderId?: number;
            /** @description Purchase Order Number supplied by the customer when the order was placed. */
            PurchaseOrderNumber?: string;
            /**
             * Format: int32
             * @description Number of items ordered of products included in the packing list for given line item
             */
            OrderedQuantity?: number;
            /**
             * Format: int32
             * @description The quantity of items in the shipment
             */
            QuantityShipped?: number;
            /**
             * Format: int32
             * @description The quantity of items on backorder
             */
            QuantityBackOrder?: number;
            /** @description Customer Line item id */
            CustomerLineItemId?: string;
            /** @description DigiKey’s Line Item number for that Sales Order */
            LineItem?: string;
            /** @description DigiKey product number for given line item */
            ProductNumber?: string;
            /** @description Product Description for given line item */
            Description?: string;
            /** @description Export control class number for given line item. See documentation from the U.S. Department of Commerce. */
            ECCN?: string;
            /** @description Customer Part Number for given line item */
            CustomerPartNumber?: string;
            /** @description Harmonized Tariff Schedule of the United States for given line item. See documentation from the U.S. International Trade Commission. */
            HTSUS?: string;
            /** @description Manufacturer part number for given line item */
            ManufacturerProductNumber?: string;
            /** @description Manufacturer Name for given line item */
            Manufacturer?: string;
            /** @description Country of origin for given line item */
            CountryOfOrigin?: string;
            /**
             * Format: int32
             * @description Quantity of products in the lot for given line item
             */
            LotQuantity?: number;
            /** @description Date Code for given line item. Format is MMM-YYYY */
            DateCode?: string;
            /** @description Lot Code for given line item */
            LotCode?: string;
        };
        PoNumberPackList: {
            /** @description List of sales orders associated with given Purchase Order Number. */
            SalesOrderPackLists?: components["schemas"]["SalesOrderPackList"][];
            /** @description Customers provided Purchase Order Number. */
            PoNumber?: string;
        };
        /** @description PackList from provided Sales Order Id */
        SalesOrderPackList: {
            /** @description Invoices from the sales order id provided */
            PackingListInvoice?: components["schemas"]["InvoicePackingList"][];
            /**
             * Format: int32
             * @description The SalesOrderID that you entered in the reques
             */
            SalesOrderIdEntered?: number;
        };
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}
export type $defs = Record<string, never>;
export interface operations {
    GetPackingList: {
        parameters: {
            query?: {
                /** @description If you would like to have a pdf for this packing list returned, set the value to “True”. */
                includePdf?: boolean;
            };
            header: {
                /** @description The Client Id for your app. */
                "X-DIGIKEY-Client-Id": string;
            };
            path: {
                /** @description Enter the Invoice ID that is to be returned. */
                invoiceId: number;
            };
            cookie?: never;
        };
        requestBody?: never;
        responses: {
            /** @description Success */
            200: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    "text/plain": components["schemas"]["InvoicePackingList"];
                    "application/json": components["schemas"]["InvoicePackingList"];
                    "text/json": components["schemas"]["InvoicePackingList"];
                };
            };
            /** @description Bad Request - The input model is invalid or malformed. */
            400: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    "text/plain": components["schemas"]["DKProblemDetails"];
                    "application/json": components["schemas"]["DKProblemDetails"];
                    "text/json": components["schemas"]["DKProblemDetails"];
                };
            };
            /** @description Unauthorized - Token is expired or Client Id or subscription are invalid. */
            401: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    "text/plain": components["schemas"]["DKProblemDetails"];
                    "application/json": components["schemas"]["DKProblemDetails"];
                    "text/json": components["schemas"]["DKProblemDetails"];
                };
            };
            /** @description The target resource is forbidden. */
            403: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    "text/plain": components["schemas"]["DKProblemDetails"];
                    "application/json": components["schemas"]["DKProblemDetails"];
                    "text/json": components["schemas"]["DKProblemDetails"];
                };
            };
            /** @description Not Found - The requested resource could not be found */
            404: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    "text/plain": components["schemas"]["DKProblemDetails"];
                    "application/json": components["schemas"]["DKProblemDetails"];
                    "text/json": components["schemas"]["DKProblemDetails"];
                };
            };
            /** @description Too Many Requests - Your rate limit has been exceeded */
            429: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    "text/plain": components["schemas"]["DKProblemDetails"];
                    "application/json": components["schemas"]["DKProblemDetails"];
                    "text/json": components["schemas"]["DKProblemDetails"];
                };
            };
            /** @description An unhandled error occurred. */
            500: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    "text/plain": components["schemas"]["DKProblemDetails"];
                    "application/json": components["schemas"]["DKProblemDetails"];
                    "text/json": components["schemas"]["DKProblemDetails"];
                };
            };
            /** @description Service Unavailable - The service is temporarily unavailable */
            503: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    "text/plain": components["schemas"]["DKProblemDetails"];
                    "application/json": components["schemas"]["DKProblemDetails"];
                    "text/json": components["schemas"]["DKProblemDetails"];
                };
            };
        };
    };
// ...
}